# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Talib MR - University Student Platform for Mauritania
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /universities:
    get:
      operationId: listUniversities
      tags: [universities]
      summary: List all universities
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [community_created, verified, official_partner]
      responses:
        "200":
          description: Universities list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/University"
                required: [success, data]
  /universities/{universityId}/faculties:
    get:
      operationId: listFaculties
      tags: [universities]
      summary: List faculties for a university
      parameters:
        - name: universityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Faculties list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Faculty"
                required: [success, data]
  /faculties/{facultyId}/departments:
    get:
      operationId: listDepartments
      tags: [universities]
      summary: List departments for a faculty
      parameters:
        - name: facultyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Departments list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Department"
                required: [success, data]
  /departments/{departmentId}/levels:
    get:
      operationId: listLevels
      tags: [universities]
      summary: List levels for a department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Levels list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Level"
                required: [success, data]
  /levels/{levelId}/groups:
    get:
      operationId: listGroups
      tags: [universities]
      summary: List groups for a level
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Group"
                required: [success, data]

  # ─── PROFILE ─────────────────────────────────────────────────────────────────
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
                required: [success, data]

    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
                required: [success, data]
  /profile/complete-onboarding:
    post:
      operationId: completeOnboarding
      tags: [profile]
      summary: Complete onboarding with academic selections
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
                required: [success, data]

  # ─── COURSES ─────────────────────────────────────────────────────────────────
...
  /subscriptions/me:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user's active subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Subscription"
                required: [success, data]
  /subscriptions/redeem:
    post:
      operationId: redeemActivationCode
      tags: [subscriptions]
      summary: Redeem an activation code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
              required: [code]
      responses:
        "200":
          description: Code redeemed, subscription activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Subscription"
                required: [success, data]
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/payment:
    post:
      operationId: submitPaymentProof
      tags: [subscriptions]
      summary: Submit manual payment proof
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "201":
          description: Payment submitted for review
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Payment"
                required: [success, data]
  /subscriptions/payments:
    get:
      operationId: listMyPayments
      tags: [subscriptions]
      summary: List current user's payment history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Payment"
                required: [success, data]

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
    Profile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        universityId:
          type: string
          nullable: true
        facultyId:
          type: string
          nullable: true
        departmentId:
          type: string
          nullable: true
        levelId:
          type: string
          nullable: true
        groupId:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        privacy:
          type: string
          enum: [private, same_university, public]
        language:
          type: string
          enum: [ar, fr, en]
        university:
          $ref: "#/components/schemas/University"
          nullable: true
        faculty:
          $ref: "#/components/schemas/Faculty"
          nullable: true
        department:
          $ref: "#/components/schemas/Department"
          nullable: true
        level:
          $ref: "#/components/schemas/Level"
          nullable: true
        group:
          $ref: "#/components/schemas/Group"
          nullable: true
        onboardingComplete:
          type: boolean
      required: [id, userId, privacy, language, onboardingComplete]
    UpdateProfileInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        privacy:
          type: string
          enum: [private, same_university, public]
        language:
          type: string
          enum: [ar, fr, en]
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        portfolioUrl:
          type: string
    OnboardingInput:
      type: object
      properties:
        universityId:
          type: string
        facultyId:
          type: string
        departmentId:
          type: string
        levelId:
          type: string
        groupId:
          type: string
        language:
          type: string
          enum: [ar, fr, en]
      required: [universityId, facultyId, departmentId, levelId]

    # ─── ACADEMIC SCHEMAS ─────────────────────────────────────────────────────
    University:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
        city:
          type: string
        logoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [community_created, verified, official_partner]
        facultyCount:
          type: integer
          nullable: true
      required: [id, name, city, status]
    Faculty:
      type: object
      properties:
        id:
          type: string
        universityId:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
      required: [id, universityId, name]
    Department:
      type: object
      properties:
        id:
          type: string
        facultyId:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
      required: [id, facultyId, name]
    Level:
      type: object
      properties:
        id:
          type: string
        departmentId:
          type: string
        name:
          type: string
        yearNumber:
          type: integer
      required: [id, departmentId, name, yearNumber]
    Group:
      type: object
      properties:
        id:
          type: string
        levelId:
          type: string
        name:
          type: string
      required: [id, levelId, name]

    # ─── COURSE SCHEMAS ───────────────────────────────────────────────────────
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled]
        source:
          type: string
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        daysRemaining:
          type: integer
      required: [id, userId, planId, planName, status, source, startsAt, expiresAt, daysRemaining]
    PaymentProofInput:
      type: object
      properties:
        planId:
          type: string
        amountMru:
          type: integer
        method:
          type: string
          enum: [bankily, masrvi, sedad, cash_agent]
        phoneNumber:
          type: string
        transactionRef:
          type: string
          nullable: true
        proofUrl:
          type: string
          nullable: true
      required: [planId, amountMru, method, phoneNumber]
    Payment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        planId:
          type: string
        planName:
          type: string
        amountMru:
          type: integer
        method:
          type: string
        status:
          type: string
          enum: [pending, under_review, approved, rejected, expired, refunded]
        phoneNumber:
          type: string
          nullable: true
        transactionRef:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, planId, planName, amountMru, method, status, createdAt]

    # ─── NOTIFICATION SCHEMAS ─────────────────────────────────────────────────
...